pyproject.toml
[tool.poetry]
name = "diss"
version = "0.1.0"
description = "Demonstration Informed Specification Search"
authors = ["Marcell Vazquez-Chanlatte <mvc@linux.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
attrs = "^21.0"
numpy = "^1.21.2"
scipy = "^1.7.1"
funcy = "^1.16"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-flake8 = "^1.0.7"
pydot = "^1.4.2"
mypy = "^0.910"
dfa = "^2.2.1"
dfa-identify = "^3.0.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
